old/unified-sort_legacy/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-sort"
version = "1.0.0"
description = "Streamlit 기반 이미지 품질 분석 및 분류 도구"
readme = "README.md"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "opencv-python>=4.8.0",
  "pillow>=10.0.0",
  "plotly>=5.18.0",
  "streamlit>=1.35.0",
  "tqdm>=4.65.0"
]

[project.optional-dependencies]
heic = ["pillow-heif>=0.15.0"]
raw = ["rawpy>=0.18.0", "imageio>=2.34.0"]
deeplearn = ["torch>=2.2.0", "torchvision>=0.17.0"]

# src 레이아웃을 사용하므로 where=["src"]
[tool.setuptools.packages.find]
where = ["src"]
include = ["unified_sort*"]
